BackEnd/pyproject.toml
[project]
name = "jaaybaanbackend"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=0.21.0",
    "django-treebeard>=4.7.0",
    "pillow>=10.0.0",
    "django-filter>=24.3",
    "django-cors-headers>=4.4.0",
]
